Add lidar driver (#5)
Contains the following:

- A skeleton lidar driver that can retrieve points from the lidar
- The start of a real PointCloud type (currently unused)

The lidar driver will currently just check "is there a wall" (are there at least 20 points in the front FOV within 20 cm), and publish a single character "planner command" if so. This will need to be forwarded over the Arduino bridge by the planner. This functionality is for construction check demo purposes only and should be deleted very soon.