Use an iterative localization algorithm
ICP tries to match the observed point cloud with the expected point cloud based on the most recent pose estimate, which can vary significantly. Instead of just accepting the first localized pose as truth, we now repeat this algorithm with the most recent estimate as the best guess for finding ICP model points. We repeat this until we either exceed the maximum number of iterations or the poses have converged.